# internal/core_backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A book management service API in Go using Gin framework..",
        "title": "Phygital Core Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://fygito.com//support",
            "email": "support@kardialab.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/digital-asset": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Digital Assets Or By Org Tag Name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digital-asset"
                ],
                "summary": "Get All Digital Assets Or By Org Tag Name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organization Tag Name Query",
                        "name": "org_tag_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/presenter.ListDigitalAssetsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/digital-asset/sync-metadata": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Sync Digital Assets Metadata (As metadata format can change so use this API to update the correct one to the DB)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digital-asset"
                ],
                "summary": "Sync Digital Assets Metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/digital-asset/{org_tag_name}/{token_id}": {
            "get": {
                "description": "Get Digital Asset Metadata By Org tag name and token ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "digital-asset"
                ],
                "summary": "Get Digital Asset Metadata By Org tag name and token ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "organization tag name",
                        "name": "org_tag_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "token_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "$ref": "#/definitions/entity.Metadata"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.Metadata": {
            "type": "object",
            "properties": {
                "animation_url": {
                    "type": "string"
                },
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MetadataAttribute"
                    }
                },
                "description": {
                    "type": "string"
                },
                "external_url": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.MetadataAttribute": {
            "type": "object",
            "properties": {
                "display_type": {
                    "type": "string"
                },
                "trait_type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "handler.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "result": {},
                "x-request-id": {
                    "type": "string"
                }
            }
        },
...
        "presenter.DigitalAssetResponse": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string"
                },
                "chain_id": {
                    "type": "integer"
                },
                "collection_id": {
                    "type": "string"
                },
                "collection_name": {
                    "type": "string"
                },
                "contract_address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {},
                "owner_email": {
                    "type": "string"
                },
                "owner_full_name": {
                    "type": "string"
                },
                "standard": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token_id": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.ListDigitalAssetsResponse": {
            "type": "object",
            "properties": {
                "digital_assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.DigitalAssetResponse"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}